requirements.txt
# Web framework
Flask==3.0.0

# API clients
withings-api==2.4.0
garminconnect==0.2.13

# Configuration
python-dotenv==1.0.0

# Optional: for better logging and debugging
colorlog==6.8.0
